# api/local_openapi.yaml
openapi: 3.1.1
info:
  title: SignalZero Symbol Store API
  version: 1.0.0
  description: |
    A minimal key/value API for persisting and retrieving SignalZero symbols by ID, domain, or tag.
servers:
  - url: https://127.0.0.1/
paths:
  /query:
    get:
      summary: Query the system by natural language or symbolic payload.
      operationId: query
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          description: Natural language query or symbolic payload in JSON
  /symbols:
    get:
      summary: Retrieve symbols by domain or tag
      operationId: querySymbols
      parameters:
        - name: symbol_domain
          in: query
          required: false
          schema:
            type: string
          description: Filter symbols by domain (e.g., root, diagnostics)
        - name: symbol_tag
          in: query
          required: false
          schema:
            type: string
          description: Filter symbols by tag (e.g., system, ritual)
        - name: start_index
          in: query
          required: false
          schema:
            type: integer
          description: Start index for pagination
        - name: limit
          in: query
          required: false
          default: 20
          maximum: 50
          schema:
            type: integer
          description: Maximum number of symbols to return
      responses:
        "200":
          description: List of matching symbols
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Symbol"
        "400":
          description: Invalid query parameters
  /symbol/{id}:
    get:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      operationId: getSymbolById
      summary: Retrieve symbol by ID
      x-openai-isConsequential: false
      responses:
        "200":
          description: Symbol retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Symbol"
        "404":
          description: Symbol not found
  /inject/{symbol_id}:
    put:
      parameters:
        - name: symbol_id
          in: path
          required: true
          schema:
            type: string
      operationId: putSymbolById
      summary: Store or replace symbol by ID
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Symbol"
      responses:
        "200":
          description: Symbol stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
    /inject/symbols:
      put:
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Symbol"
        responses:
          "200":
            description: Symbol stored
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string

      
    /domains:
      get:
        summary: List all available symbol domains
        operationId: listDomains
        responses:
          "200":
            description: List of symbol domains
            content:
              application/json:
                schema:
                  type: array
                  items:
                    type: string
          "500":
            description: Error retrieving domain list
          
components:
  schemas:
    Symbol:
      type: object
      required:
        - id
        - triad
        - macro
        - facets
      properties:
        id:
          type: string
          example: SZ:BOOT-SEAL-001
        name:
          type: string
          example: Boot Seal
        macro:
          type: string
          example: seal(boot) → archive(pre_anchor, post_anchor) → confirm(invariants)
        gate:
          type: array
          items:
            type: string
          example:
            - integrity
        facets:
          type: object
          required:
            - function
            - topology
            - commit
            - gate
            - substrate
            - temporal
            - invariants
          properties:
            function:
              type: string
              example: govern
            topology:
              type: string
              example: linear
            commit:
              type: string
              example: ledger
            gate:
              type: array
              items:
                type: string
              example:
                - integrity
            substrate:
              type: array
              items:
                type: string
              example:
                - symbolic
            temporal:
              type: string
              example: instant
            invariants:
              type: array
              items:
                type: string
              example:
                - non-coercion
                - auditability
                - baseline-integrity
        failure_mode:
          type: string
          example: Seal rejected if invariants broken
        linked_patterns:
          type: array
          items:
            type: string
          example:
            - SZ:STB-Signal-Anchor-006
            - SZ:DIA-Recursive-Unbind-046
        symbolic_role:
          type: string
          example: Boot Seal
        triad:
          type: string
          example: ⟐⇌Σ
        invocations:
          type: array
          items:
            type: string
          example:
            - Seal this boot.
        symbol_domain:
          type: string
          example: root
        symbol_tag:
          type: string
          example: system
        version:
          type: integer
          example: 1
        origin:
          type: string
          example: boot-sequence
        scope:
          type: array
          items:
            type: string
          example:
            - post-commit
            - archive